[WEAV-159] Color Extension 모듈 위치 변경
