+++ Projects/DesignSystem/DesignCore/Sources/Color+Ext.swift
@@ -1,0 +1,32 @@
+//
+//  Color+Ext.swift
+//  DesignCore
+//
+//  Created by 김지수 on 2/24/25.
+//  Copyright © 2025 com.weave. All rights reserved.
+//
+
+import SwiftUI
+import UIKit
+
+extension Color {
+    public init(hex: Int, opacity: Double = 1.0) {
+        let red = Double((hex >> 16) & 0xff) / 255
+        let green = Double((hex >> 8) & 0xff) / 255
+        let blue = Double((hex >> 0) & 0xff) / 255
+
+        self.init(.sRGB, red: red, green: green, blue: blue, opacity: opacity)
+    }
+}
+
+
+extension UIColor {
+    public convenience init(hex: Int, alpha: CGFloat = 1.0) {
+        self.init(
+            red: CGFloat((hex >> 16) & 0xFF) / 255.0,
+            green: CGFloat((hex >> 8) & 0xFF) / 255.0,
+            blue: CGFloat(hex & 0xFF) / 255.0,
+            alpha: alpha
+        )
+    }
+}
