+++ Projects/DesignSystem/DesignCore/Sources/TagListView/TagListCollectionViewCell.swift
@@ -71,14 +71,3 @@ class TagListCollectionViewCell: UICollectionViewCell {
         }
     }
 }
-
-extension UIColor {
-    convenience init(hex: Int, alpha: CGFloat = 1.0) {
-        self.init(
-            red: CGFloat((hex >> 16) & 0xFF) / 255.0,
-            green: CGFloat((hex >> 8) & 0xFF) / 255.0,
-            blue: CGFloat(hex & 0xFF) / 255.0,
-            alpha: alpha
-        )
-    }
-}
